Fix SecurityActions issues
